requirements.txt
PyQt5>=5.15.0
requests>=2.25.1
configparser>=5.0.2
PyPDF2>=3.0.0  
pdfminer.six>=20211012  